Good zoom in and out starting point.
Everything seems to work fine however there are
some weird quirks or "bugs". Here is everything I've personally
noticed:

1. When zoomed out the elements seem to have some sort of acceleration
and are very slow to move.

2. The elements seem to change size and shape depending on where
on the grid they're located.

3. This is very minor "quirk" but the zoom in does not zoom in where
the mouse is actually pointing. Just generally zooms into
the top middle part of the screen.

More issues might arise.

However in terms of actual bugs or weird behaviour for the elements
or the grid, everything seems inorder and this is a great
starting point for zoom in functionality.